+++ R/tbl-sql.r
@@ -60,7 +60,7 @@ as.data.frame.tbl_sql <- function(x, row.names = NULL, optional = NULL,
 
 #' @export
 print.tbl_sql <- function(x, ..., n = NULL, width = NULL) {
-  cat("Source:     query ", dim_desc(x), "\n", sep = "")
+  cat("Source:     ", tbl_desc(x), "\n", sep = "")
   cat("Database:   ", src_desc(x$src), "\n", sep = "")
 
   grps <- op_grps(x$ops)
@@ -77,6 +77,8 @@ print.tbl_sql <- function(x, ..., n = NULL, width = NULL) {
   print(trunc_mat(x, n = n, width = width))
   invisible(x)
 }
+
+
 
 #' @export
 dimnames.tbl_sql <- function(x) {