+++ tests/testthat/test-output.R
@@ -4,9 +4,10 @@ test_that("ungrouped output", {
   if (packageVersion("tibble") < "1.0-10")
     skip("need tibble 1.0-10 or later for this test")
 
-  mtcars_mem <- src_memdb() %>% copy_to(mtcars, name = random_table_name())
+  mtcars_mem <- src_memdb() %>%
+    copy_to(mtcars, name = "mtcars-output-test", overwrite = TRUE)
   iris_mem <- src_memdb() %>%
-    copy_to(iris, name = random_table_name()) %>%
+    copy_to(iris, name = "iris-output-test", overwrite = TRUE) %>%
     group_by(Species) %>%
     arrange(Sepal.Length)